requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.1
python-multipart==0.0.9
jinja2==3.1.3

# AI - Gemini + LangGraph + Observability
google-genai>=0.6.0
langgraph>=0.2.27
langfuse>=2.31.0
python-dotenv==1.0.1

# Optional Cloud Storage (keys present in env but not required by app flow)
azure-storage-blob>=12.19.0

# Utilities
requests==2.32.3
numpy==1.26.4
